requirements.txt
streamlit>=1.26.0
streamlit-drawable-canvas
numpy
joblib
tensorflow
pandas
matplotlib
scikit-learn
Pillow
torch
transformers
opencv-python
requests
nltk